Refactored Playboardscreen by extracting most of the code
