+++ app/src/main/java/at/aau/serg/websocketbrokerdemo/ui/PlayboardScreen.kt
@@ -1,91 +1,61 @@
 package at.aau.serg.websocketbrokerdemo.ui
 
 import android.widget.Toast
-import androidx.compose.animation.animateColorAsState
-import androidx.compose.animation.core.FastOutSlowInEasing
-import androidx.compose.animation.core.animateFloatAsState
-import androidx.compose.animation.core.tween
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.rotate
-import androidx.compose.ui.draw.scale
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import at.aau.serg.websocketbrokerdemo.data.PlayerMoney
 import kotlinx.coroutines.delay
 import at.aau.serg.websocketbrokerdemo.data.properties.Property
 import at.aau.serg.websocketbrokerdemo.data.properties.PropertyViewModel
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.horizontalScroll
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.widthIn
-import androidx.compose.foundation.layout.wrapContentHeight
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.mutableFloatStateOf
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.draw.alpha
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.text.TextStyle
-import at.aau.serg.websocketbrokerdemo.data.properties.getDrawableIdFromName
 import at.aau.serg.websocketbrokerdemo.GameWebSocketClient
 import at.aau.serg.websocketbrokerdemo.data.ChatEntry
-import at.aau.serg.websocketbrokerdemo.data.properties.PropertyColor
 import androidx.compose.ui.text.input.KeyboardType
 import at.aau.serg.websocketbrokerdemo.data.CheatEntry
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.runtime.snapshots.SnapshotStateList
-import androidx.compose.ui.text.font.FontFamily
 import at.aau.serg.websocketbrokerdemo.data.messages.DealProposalMessage
 import at.aau.serg.websocketbrokerdemo.data.messages.DealResponseMessage
 import at.aau.serg.websocketbrokerdemo.data.messages.DealResponseType
 import com.google.gson.Gson
 import at.aau.serg.websocketbrokerdemo.logic.ShakeDetector
 import at.aau.serg.websocketbrokerdemo.data.messages.ShakeMessage
+import at.aau.serg.websocketbrokerdemo.ui.components.DiceRollingButton
+import at.aau.serg.websocketbrokerdemo.ui.components.PlayerCard
+import at.aau.serg.websocketbrokerdemo.ui.components.dialogs.ActionMenuDialog
+import at.aau.serg.websocketbrokerdemo.ui.components.alerts.CardPopup
+import at.aau.serg.websocketbrokerdemo.ui.components.alerts.ChatOverlay
+import at.aau.serg.websocketbrokerdemo.ui.components.alerts.CheatTerminalOverlay
+import at.aau.serg.websocketbrokerdemo.ui.components.alerts.PassedGoAlertBox
+import at.aau.serg.websocketbrokerdemo.ui.components.alerts.PropertyPopup
+import at.aau.serg.websocketbrokerdemo.ui.components.alerts.TaxPaymentAlertBox
 
 
 fun extractPlayerId(message: String): String {
@@ -172,7 +142,6 @@ fun PlayboardScreen(
     var ownedProperties by remember { mutableStateOf<List<Property>>(emptyList()) }
 
     var showActionMenu by remember { mutableStateOf(false) }
-
 
 
     // ShakeDetector:
@@ -431,65 +400,13 @@ fun PlayboardScreen(
 
         // ACTION MENU
         if (showActionMenu) {
-            AlertDialog(
-                onDismissRequest = { showActionMenu = false },
-                confirmButton = {},
-                dismissButton = {},
-                text = {
-                    Column(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(16.dp),
-                        verticalArrangement = Arrangement.spacedBy(16.dp),
-                        horizontalAlignment = Alignment.CenterHorizontally
-                    ) {
-                        if (isMyTurn && !turnEnded) {
-                            Button(
-                                onClick = {
-                                    showDealDialog = true
-                                    showActionMenu = false
-                                },
-                                modifier = Modifier.fillMaxWidth(),
-                                colors = ButtonDefaults.buttonColors(containerColor = Color(0xFF9C27B0)) // Violett
-                            ) {
-                                Text("Start Deal", color = Color.White)
-                            }
-                        }
-
-                        Button(
-                            onClick = {
-                                onBackToLobby()
-                                showActionMenu = false
-                            },
-                            modifier = Modifier.fillMaxWidth(),
-                            colors = ButtonDefaults.buttonColors(containerColor = Color(0xFF0074cc)) // Blau
-                        ) {
-                            Text("Back to Lobby", color = Color.White)
-                        }
-
-                        Button(
-                            onClick = {
-                                onGiveUp()
-                                showActionMenu = false
-                            },
-                            enabled = isMyTurn,
-                            colors = ButtonDefaults.buttonColors(
-                                containerColor = if (isMyTurn) Color.Red else Color.Gray
-                            ),
-                            modifier = Modifier.fillMaxWidth()
-                        ) {
-                            Text("Give Up", color = Color.White)
-                        }
-
-                        Button(
-                            onClick = { showActionMenu = false },
-                            modifier = Modifier.fillMaxWidth(),
-                            colors = ButtonDefaults.buttonColors(containerColor = Color.LightGray)
-                        ) {
-                            Text("Close", color = Color.Black)
-                        }
-                    }
-                }
+            ActionMenuDialog(
+                isMyTurn = isMyTurn,
+                turnEnded = turnEnded,
+                onDismiss = { showActionMenu = false },
+                onStartDeal = { showDealDialog = true },
+                onBackToLobby = onBackToLobby,
+                onGiveUp = onGiveUp
             )
         }
 
@@ -503,915 +420,211 @@ fun PlayboardScreen(
                     canBuy = false
                 }
             }
-            val imageResId = getDrawableIdFromName(selectedProperty!!.image, context)
 
-            AlertDialog(
-                modifier = Modifier
-                    .width(300.dp)
-                    .height(400.dp),
-                onDismissRequest = {
+            PropertyPopup(
+                selectedProperty = selectedProperty!!,
+                canBuy = canBuy,
+                isMyTurn = currentPlayerId == localPlayerId,
+                openedByClick = openedByClick,
+                localPlayerId = localPlayerId,
+                gameEvents = gameEvents,
+                webSocketClient = webSocketClient,
+                onDismiss = {
                     selectedProperty = null
                     openedByClick = false
                     canBuy = false
-                },
-                title = {
-                    Box(
-                        modifier = Modifier.fillMaxWidth(),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        Text(
-                            text = selectedProperty!!.name,
-                            style = TextStyle(
-                                fontSize = 20.sp,
-                                fontWeight = FontWeight.Bold
-                            )
-                        )
-                    }
-                },
-                text = {
-                    Column(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .height(300.dp),
-                        verticalArrangement = Arrangement.Center,
-                        horizontalAlignment = Alignment.CenterHorizontally
-                    ) {
-                        if (imageResId != 0) {
-                            Box(
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .height(250.dp),
-                                contentAlignment = Alignment.Center
-                            ) {
-                                Image(
-                                    painter = painterResource(imageResId),
-                                    contentDescription = selectedProperty!!.name,
-                                    contentScale = ContentScale.Fit,
-                                    modifier = Modifier
-                                        .width(180.dp)
-                                        .height(240.dp)
-                                )
-                            }
-                        }
-                    }
-                },
-                confirmButton = {
-                    Row(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(8.dp),
-                        horizontalArrangement = Arrangement.SpaceEvenly
-                    ) {
-                        Button(
-                            onClick = {
-                                Toast.makeText(context, "❌ Purchase declined", Toast.LENGTH_SHORT).show()
-                                gameEvents.add("❌ Purchase declined")
-                                selectedProperty = null
-                                openedByClick = false
-                                canBuy = false
-                            },
-                            colors = ButtonDefaults.buttonColors(containerColor = Color(0xFF0074cc))
-                        ) {
-                            Text("Exit")
-                        }
-                        if (canBuy && localPlayerId == currentPlayerId) {
-                            Button(
-                                onClick = {
-                                    val propertyName = selectedProperty?.name ?: "a property"
-                                    val message = "Purchase confirmed ✅ : $propertyName"
-
-                                    webSocketClient.sendMessage("BUY_PROPERTY:${selectedProperty?.id}")
-
-                                    Toast.makeText(context, message, Toast.LENGTH_LONG).show()
-                                    gameEvents.add(message)
-
-                                    selectedProperty = null
-                                    openedByClick = false
-                                    canBuy = false
-                                },
-                                colors = ButtonDefaults.buttonColors(
-                                    containerColor = Color(
-                                        0xFF0074cc
-                                    )
-                                )
-                            ) {
-                                Text("Buy")
-                            }
-                        }
-                        // Add Pay Rent button if property is owned by another player
-                        if (selectedProperty?.ownerId != null &&
-                            selectedProperty?.ownerId != localPlayerId &&
-                            localPlayerId == currentPlayerId
-                        ) {
-                            Button(
-                                onClick = {
-                                    val msg = "💸 Rent paid for ${selectedProperty?.name ?: "a property"}"
-                                    Toast.makeText(context, msg, Toast.LENGTH_SHORT).show()
-                                    gameEvents.add(msg)
-
-                                    webSocketClient.logic().payRent(selectedProperty?.id ?: -1)
-                                    rentPaid = true
-                                    selectedProperty = null
-                                    openedByClick = false
-                                    canBuy = false
-                                },
-                                enabled = !rentPaid,
-                                colors = ButtonDefaults.buttonColors(
-                                    containerColor = Color(0xFFE57373),
-                                    disabledContainerColor = Color(0xFFBDBDBD)
-                                )
-                            ) {
-                                Text("Pay Rent")
-                            }
-                        }
-                    }
-                },
-                dismissButton = {}
+                }
             )
         }
 
         // Passed GO Alert
         if (showPassedGoAlert) {
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .fillMaxHeight(0.3f)
-                    .background(Color(0xFF4CAF50).copy(alpha = 0.9f))
-                    .padding(16.dp),
-                contentAlignment = Alignment.Center
-            ) {
-                Column(
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    verticalArrangement = Arrangement.Center
-                ) {
-                    Text(
-                        text = "Glückwunsch!",
-                        style = TextStyle(
-                            fontSize = 28.sp,
-                            fontWeight = FontWeight.Bold,
-                            color = Color.White
-                        )
-                    )
-                    Spacer(modifier = Modifier.height(8.dp))
-                    Text(
-                        text = "$passedGoPlayerName fuhr über los und erhält 200€!",
-                        style = TextStyle(
-                            fontSize = 20.sp,
-                            color = Color.White
-                        )
-                    )
-                }
-            }
+            PassedGoAlertBox(playerName = passedGoPlayerName)
         }
 
         // Tax Payment Alert
         if (showTaxPaymentAlert) {
-            Box(
+            TaxPaymentAlertBox(
+                playerName = taxPaymentPlayerName,
+                amount = taxPaymentAmount,
+                taxType = taxPaymentType
+            )
+        }
+
+        Box(modifier = Modifier.fillMaxSize()) {
+            Column(
                 modifier = Modifier
-                    .fillMaxWidth()
-                    .fillMaxHeight(0.3f)
-                    .background(Color(0xFFFFA500).copy(alpha = 0.9f))
+                    .align(Alignment.BottomStart)
                     .padding(16.dp),
-                contentAlignment = Alignment.Center
-            ) {
-                Column(
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    verticalArrangement = Arrangement.Center
-                ) {
-                    Text(
-                        text = "Autsch!",
-                        style = TextStyle(
-                            fontSize = 28.sp,
-                            fontWeight = FontWeight.Bold,
-                            color = Color.White
-                        )
-                    )
-                    Spacer(modifier = Modifier.height(8.dp))
-                    Text(
-                        text = "$taxPaymentPlayerName muss ${taxPaymentAmount}€ $taxPaymentType an die Bank zahlen!",
-                        style = TextStyle(
-                            fontSize = 20.sp,
-                            color = Color.White
-                        )
-                    )
-                }
-            }
-        }
-    }
-
-    Box(modifier = Modifier.fillMaxSize()) {
-        Column(
-            modifier = Modifier
-                .align(Alignment.BottomStart)
-                .padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(8.dp)
-        ) {
-            // Cheat-terminal toggle
-            Button(
-                onClick = { cheatTerminalOpen = !cheatTerminalOpen },
-                colors = ButtonDefaults.buttonColors(containerColor = Color(0xFF3933cc)),
-                modifier = Modifier
-                    .width(140.dp)
-                    .height(48.dp)
-            ) {
-                Text(
-                    if (cheatTerminalOpen) "Close Terminal" else "Open Terminal",
-                    fontSize = 12.sp
-                )
-            }
-
-            // Chat toggle
-            Button(
-                onClick = { chatOpen = !chatOpen },
-                colors = ButtonDefaults.buttonColors(containerColor = Color(0xFF0074cc)),
-                modifier = Modifier
-                    .width(140.dp)
-                    .height(48.dp)
-            ) {
-                Text(
-                    if (chatOpen) "Close Chat" else "Open Chat",
-                    fontSize = 14.sp
-                )
-            }
-        }
-    }
-
-    // Chat Overlay
-
-    if (chatOpen) {
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(Color.Black.copy(alpha = 0.3f)) // halbtransparenter schwarzer Hintergrund
-                .padding(32.dp)
-        ) {
-            IconButton(
-                onClick = { chatOpen = false },
-                modifier = Modifier.align(Alignment.TopStart)
-            ) {
-                Icon(Icons.Filled.ArrowBack, contentDescription = "Back", tint = Color.Black)
-            }
-            Column(
-                modifier = Modifier
-                    .align(Alignment.Center)
-                    .fillMaxWidth()
-                    .background(Color.White.copy(alpha = 0.85f), RoundedCornerShape(12.dp))
-                    .padding(16.dp)
-            ) {
-                // Nachrichtenliste
-                LazyColumn(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .weight(1f)
-                        .padding(8.dp),
-                    reverseLayout = true
-                ) {
-                    items(chatMessages.reversed()) { entry ->
-                        val isOwnMessage = entry.senderId == currentPlayerId
-
-                        Row(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .padding(vertical = 4.dp),
-                            horizontalArrangement = if (isOwnMessage) Arrangement.End else Arrangement.Start
-                        ) {
-                            Box(
-                                modifier = Modifier
-                                    .background(
-                                        if (isOwnMessage) Color(0xFFDCF8C6) else Color.White,
-                                        RoundedCornerShape(12.dp)
-                                    )
-                                    .padding(12.dp)
-                                    .widthIn(max = 240.dp)
-                            ) {
-                                Column {
-                                    val nameColor = playerColorMap[entry.senderId] ?: Color.Gray
-                                    Text(
-                                        text = entry.senderName,
-                                        fontWeight = FontWeight.Bold,
-                                        fontSize = 12.sp,
-                                        color = nameColor
-                                    )
-                                    Spacer(modifier = Modifier.height(4.dp))
-                                    Text(
-                                        text = entry.message,
-                                        color = Color.Black,
-                                        fontSize = 16.sp
-                                    )
-                                }
-                            }
-                        }
-                    }
-
-                }
-
-                Spacer(modifier = Modifier.height(8.dp))
-
-                // Eingabe und Senden
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    modifier = Modifier.fillMaxWidth()
-                ) {
-                    TextField(
-                        value = chatInput,
-                        onValueChange = { chatInput = it },
-                        modifier = Modifier.weight(1f),
-                        placeholder = { Text("Type your message...") }
-                    )
-                    Spacer(modifier = Modifier.width(8.dp))
-                    Button(
-                        onClick = {
-                            if (chatInput.isNotBlank()) {
-                                webSocketClient.logic().sendChatMessage(currentPlayerId, chatInput)
-                                chatInput = "" // Nach Senden Eingabefeld leeren
-                            }
-                        }
-                    ) {
-                        Text("Send")
-                    }
-                }
-            }
-        }
-    }
-
-    // DEALS
-
-    if (showDealDialog) {
-        DealDialog(
-            players = players.filter { it.id != localPlayerId },
-            senderId = localPlayerId,
-            allProperties = properties,
-            receiver = selectedReceiver,
-            avatarMap = avatarMap,
-            onReceiverChange = { selectedReceiver = it },
-            onSendDeal = { deal ->
-                val json = Gson().toJson(deal)
-                webSocketClient.sendMessage(json)
-                showDealDialog = false
-                selectedReceiver = null
-            },
-            onDismiss = {
-                showDealDialog = false
-                selectedReceiver = null
-            }
-        )
-    }
-
-    if (showIncomingDialog && incomingDeal != null) {
-        val receiverProps = properties.filter { it.ownerId == localPlayerId }.map { it.id }
-
-        IncomingDealDialog(
-            proposal = incomingDeal,
-            senderName = players.find { it.id == incomingDeal.fromPlayerId }?.name ?: "???",
-            allProperties = properties,
-            receiverProperties = receiverProps,
-            isMyTurn = isMyTurn,
-            onAccept = {
-                val response = DealResponseMessage(
-                    type = "DEAL_RESPONSE",
-                    fromPlayerId = localPlayerId,
-                    toPlayerId = incomingDeal.fromPlayerId,
-                    responseType = DealResponseType.ACCEPT,
-                    counterPropertyIds = listOf(),
-                    counterMoney = 0
-                )
-                webSocketClient.sendMessage(Gson().toJson(response))
-                setShowIncomingDialog(false)
-                setIncomingDeal(null)
-            },
-            onDecline = {
-                val response = DealResponseMessage(
-                    type = "DEAL_RESPONSE",
-                    fromPlayerId = localPlayerId,
-                    toPlayerId = incomingDeal.fromPlayerId,
-                    responseType = DealResponseType.DECLINE,
-                    counterPropertyIds = listOf(),
-                    counterMoney = 0
-                )
-                webSocketClient.sendMessage(Gson().toJson(response))
-                setShowIncomingDialog(false)
-                setIncomingDeal(null)
-            },
-            onCounter = {
-                isCountering = true
-                setShowIncomingDialog(false)
-            }
-        )
-    }
-
-    if (isCountering && incomingDeal != null) {
-        DealDialog(
-            players = players.filter { it.id != localPlayerId },
-            senderId = localPlayerId,
-            allProperties = properties,
-            receiver = players.find { it.id == incomingDeal.fromPlayerId },
-            avatarMap = avatarMap,
-            initialRequested = incomingDeal.offeredPropertyIds,
-            initialOffered = incomingDeal.requestedPropertyIds,
-            initialMoney = incomingDeal.offeredMoney,
-            onSendDeal = { counter ->
-                val counterProposal = DealProposalMessage(
-                    type = "DEAL_PROPOSAL",
-                    fromPlayerId = localPlayerId,
-                    toPlayerId = incomingDeal.fromPlayerId,
-                    offeredPropertyIds = counter.requestedPropertyIds,
-                    requestedPropertyIds = counter.offeredPropertyIds,
-                    offeredMoney = counter.offeredMoney
-                )
-                webSocketClient.sendMessage(Gson().toJson(counterProposal))
-                isCountering = false
-                setIncomingDeal(null)
-            },
-            onDismiss = {
-                isCountering = false
-                setIncomingDeal(null)
-            }
-        )
-    }
-
-    // Cheat Terminal Overview
-    if (cheatTerminalOpen) {
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(Color.Black.copy(alpha = 0.65f))
-                .padding(32.dp)
-        ) {
-            // ← Back button in top-left
-            IconButton(
-                onClick = { cheatTerminalOpen = false },
-                modifier = Modifier.align(Alignment.TopStart)
-            ) {
-                Icon(Icons.Filled.ArrowBack, contentDescription = "Back", tint = Color(0xFFCCFF90))
-            }
-            Column(
-                modifier = Modifier
-                    .align(Alignment.Center)
-                    .fillMaxWidth()
-                    .background(Color.Black.copy(alpha = 0.65f), RoundedCornerShape(12.dp))
-                    .padding(16.dp)
-            ) {
-                // Nachrichtenliste
-                LazyColumn(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .weight(1f)
-                        .padding(8.dp),
-                    reverseLayout = true
-                ) {
-                    items(cheatMessages.reversed()) { entry ->
-                        Text(
-                            text = "${entry.senderName.lowercase()}@monopoly > ${entry.message}",
-                            fontFamily = FontFamily.Monospace,
-                            fontSize = 14.sp,
-                            color = Color(0xFF00FF00)
-                        )
-                    }
-                }
-
-                Spacer(modifier = Modifier.height(8.dp))
-
-                // Eingabe und Senden
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    modifier = Modifier.fillMaxWidth()
-                ) {
-                    TextField(
-                        value = cheatInput,
-                        onValueChange = { cheatInput = it },
-                        modifier = Modifier.weight(1f),
-                        placeholder = { Text("Type your cheat code...") }
-                    )
-                    Spacer(modifier = Modifier.width(8.dp))
-                    Button(
-                        onClick = {
-                            if (cheatInput.isNotBlank()) {
-                                webSocketClient.logic()
-                                    .sendCheatMessage(currentPlayerId, cheatInput)
-                                cheatInput = "" // Nach Senden Eingabefeld leeren
-                            }
-                        }
-                    ) {
-                        Text("Send")
-                    }
-                }
-            }
-        }
-    }
-    // Popup for CHANCE and COMMUNITY_CHEST
-    if (drawnCardType != null && drawnCardId != null) {
-        // Build the resource name, e.g. "chance_2" or "community_chest_7"
-        val resName = "${drawnCardType.lowercase()}_${drawnCardId}"
-        val imageResId = context.resources.getIdentifier(resName, "drawable", context.packageName)
-
-        AlertDialog(
-            onDismissRequest = { onCardDialogDismiss() },
-            title = {
-                Text(
-                    text = if (drawnCardType == "CHANCE") "Ereigniskarte" else "Gemeinschaftskarte",
-                    fontSize = 20.sp,
-                    fontWeight = FontWeight.Bold
-                )
-            },
-            text = {
-                Column(
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    verticalArrangement = Arrangement.Center
-                ) {
-                    // Show the card image if it exists
-                    if (imageResId != 0) {
-                        Image(
-                            painter = painterResource(id = imageResId),
-                            contentDescription = "Card $resName",
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .height(200.dp)
-                                .clip(RoundedCornerShape(8.dp)),
-                            contentScale = ContentScale.Fit
-                        )
-                    } else {
-                        // Fallback
-                        Text(
-                            text = drawnCardDesc ?: "",
-                            fontSize = 16.sp,
-                            fontWeight = FontWeight.Medium
-                        )
-                    }
-                }
-            },
-            confirmButton = {
-                Button(
-                    onClick = { onCardDialogDismiss() }
-                ) {
-                    Text("OK")
-                }
-            },
-            dismissButton = {}
-        )
-    }
-}
-
-@Composable
-fun PlayerCard(
-    player: PlayerMoney,
-    ownedProperties: List<Property>,
-    allProperties: List<Property>,
-    isCurrentPlayer: Boolean,
-    playerIndex: Int,
-    onPropertySetClicked: (PropertyColor) -> Unit,
-    webSocketClient: GameWebSocketClient
-) {
-    val playerColors = listOf(
-        Color(0x80FF0000), // Less saturated Red
-        Color(0x800000FF), // Less saturated Blue
-        Color(0x8000FF00), // Less saturated Green
-        Color(0x80FFFF00)  // Less saturated Yellow
-    )
-
-    val backgroundColor = playerColors[playerIndex].copy(alpha = 0.4f)
-
-    var selectedColorSet by remember { mutableStateOf<PropertyColor?>(null) }
-
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .wrapContentHeight()
-            .then(
-                if (isCurrentPlayer) {
-                    Modifier.border(
-                        width = 4.dp,
-                        color = Color.Black,
-                        shape = RoundedCornerShape(8.dp)
-                    )
-                } else Modifier
-            ),
-        colors = CardDefaults.cardColors(containerColor = backgroundColor),
-        shape = RoundedCornerShape(8.dp)
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(16.dp),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = player.name,
-                    color = Color.White,
-                    fontSize = 14.sp,
-                    fontWeight = FontWeight.Bold
-                )
-                Text(
-                    text = "EUR ${player.money}",
-                    color = Color.White,
-                    fontSize = 12.sp,
-                    fontWeight = FontWeight.Bold
-                )
-                Text(
-                    text = "Pos: ${player.position}",
-                    color = Color.White,
-                    fontSize = 12.sp
-                )
-            }
-
-            Text(
-                text = "ID: ${player.id}",
-                color = Color.White,
-                fontSize = 6.sp
-            )
-
-            HorizontalDivider(
-                modifier = Modifier.padding(vertical = 4.dp),
-                thickness = 1.dp,
-                color = Color.White
-            )
-
-            BesitzkartenGrid(
-                ownedProperties = ownedProperties,
-                allProperties = allProperties,
-                onPropertySetClicked = { colorSet -> selectedColorSet = colorSet }
-            )
-
-            if (selectedColorSet != null) {
-                PropertySetPopup(
-                    colorSet = selectedColorSet!!,
-                    ownedProperties = ownedProperties.filter { getColorForPosition(it.position) == selectedColorSet },
-                    allProperties = allProperties.filter { getColorForPosition(it.position) == selectedColorSet },
-                    onDismiss = { selectedColorSet = null },
-                    onSellProperty = { propertyId ->
-                        webSocketClient.logic().sellProperty(propertyId)
-                    }
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun BesitzkartenGrid(
-    ownedProperties: List<Property>,
-    allProperties: List<Property>,
-    onPropertySetClicked: (PropertyColor) -> Unit
-) {
-    val propertySets = PropertyColor.entries.toTypedArray()
-
-    LazyVerticalGrid(
-        columns = GridCells.Fixed(5),
-        modifier = Modifier
-            .heightIn(max = 120.dp)
-    ) {
-        items(propertySets.size) { index ->
-            val colorSet = propertySets[index]
-            PropertySetCard(
-                colorSet = colorSet,
-                ownedProperties = ownedProperties,
-                allProperties = allProperties,
-                onClick = { onPropertySetClicked(colorSet) }
-            )
-        }
-    }
-}
-
-@Composable
-fun PropertySetCard(
-    colorSet: PropertyColor,
-    ownedProperties: List<Property>,
-    allProperties: List<Property>,
-    onClick: () -> Unit
-) {
-    val propertiesInSet = ownedProperties.filter { getColorForPosition(it.position) == colorSet }
-    val ownsCompleteSet = checkCompleteSet(colorSet, propertiesInSet, allProperties)
-
-    val cardAlpha = if (propertiesInSet.isEmpty()) 0.3f else if (ownsCompleteSet) 1f else 0.6f
-
-    Card(
-        modifier = Modifier
-            .padding(2.dp)
-            .aspectRatio(1f)
-            .clickable { onClick() },
-        shape = RoundedCornerShape(12.dp),
-        colors = CardDefaults.cardColors(
-            containerColor = getColorForSet(colorSet).copy(alpha = cardAlpha)
-        )
-    ) {}
-}
-
-fun checkCompleteSet(colorSet: PropertyColor, owned: List<Property>, allProperties: List<Property>): Boolean {
-    val totalInSet = allProperties.count { getColorForPosition(it.position) == colorSet }
-    return owned.size == totalInSet
-}
-
-fun getColorForSet(colorSet: PropertyColor): Color {
-    return when (colorSet) {
-        PropertyColor.BROWN -> Color(0xFF964B00)
-        PropertyColor.LIGHT_BLUE -> Color(0xFFADD8E6)
-        PropertyColor.PINK -> Color(0xFFFFC0CB)
-        PropertyColor.ORANGE -> Color(0xFFFFA500)
-        PropertyColor.RED -> Color.Red
-        PropertyColor.YELLOW -> Color.Yellow
-        PropertyColor.GREEN -> Color.Green
-        PropertyColor.DARK_BLUE -> Color(0xFF00008B)
-        PropertyColor.RAILROAD -> Color(0xFF8B4513)
-        PropertyColor.UTILITY -> Color (0xFF20B2AA)
-    }
-}
-
-fun getColorForPosition(position: Int): PropertyColor {
-    return when (position) {
-        1, 3 -> PropertyColor.BROWN
-        6, 8, 9 -> PropertyColor.LIGHT_BLUE
-        11, 13, 14 -> PropertyColor.PINK
-        16, 18, 19 -> PropertyColor.ORANGE
-        21, 23, 24 -> PropertyColor.RED
-        26, 27, 29 -> PropertyColor.YELLOW
-        31, 32, 34 -> PropertyColor.GREEN
-        37, 39 -> PropertyColor.DARK_BLUE
-        5, 15, 25, 35 -> PropertyColor.RAILROAD
-        12, 28 -> PropertyColor.UTILITY
-        else -> error("Unhandled position: $position")
-    }
-}
-
-@Composable
-fun PropertySetPopup(
-    colorSet: PropertyColor,
-    ownedProperties: List<Property>,
-    allProperties: List<Property>,
-    onDismiss: () -> Unit,
-    onSellProperty: (Int) -> Unit
-) {
-    val context = LocalContext.current
-    val propertiesInSet = allProperties.filter { getColorForPosition(it.position) == colorSet }
-
-    AlertDialog(
-        modifier = Modifier
-            .width(420.dp)
-            .height(550.dp),
-        onDismissRequest = onDismiss,
-        title = {
-            Text(
-                text = "${colorSet.name} Set",
-                style = TextStyle(
-                    fontSize = 22.sp,
-                    fontWeight = FontWeight.Bold
-                ),
-                modifier = Modifier.fillMaxWidth(),
-            )
-        },
-        text = {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(350.dp)
-                    .horizontalScroll(rememberScrollState())
-                    .padding(8.dp),
-                horizontalArrangement = Arrangement.spacedBy(12.dp)
-            ) {
-                propertiesInSet.forEach { property ->
-                    val imageResId = getDrawableIdFromName(property.image, context)
-                    val isOwned = ownedProperties.any { it.id == property.id }
-                    Box(
-                        modifier = Modifier
-                            .width(180.dp)
-                            .aspectRatio(0.7f)
-                            .clip(RoundedCornerShape(12.dp))
-                            .background(
-                                Color.LightGray.copy(alpha = 1f)
-                            ),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        if (imageResId != 0) {
-                            Image(
-                                painter = painterResource(id = imageResId),
-                                contentDescription = property.name,
-                                contentScale = ContentScale.Fit,
-                                modifier = Modifier
-                                    .fillMaxSize()
-                                    .alpha(if (isOwned) 1f else 0.4f)
-                            )
-                        } else {
-                            Text(
-                                text = property.name,
-                                color = Color.Black,
-                                fontSize = 12.sp
-                            )
-                        }
-                    }
-                }
-            }
-        },
-        confirmButton = {
-            Column(
-                modifier = Modifier.fillMaxWidth(),
                 verticalArrangement = Arrangement.spacedBy(8.dp)
             ) {
-                ownedProperties.forEach { property ->
-                    Button(
-                        onClick = {
-                            onSellProperty(property.id)
-                            onDismiss()
-                        },
-                        colors = ButtonDefaults.buttonColors(containerColor = Color(0xFFE53935)),
-                        shape = RoundedCornerShape(20.dp),
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(horizontal = 8.dp)
-                    ) {
-                        Text("Sell ${property.name}", color = Color.White)
-                    }
-                }
+                // Cheat-terminal toggle
                 Button(
-                    onClick = onDismiss,
-                    colors = ButtonDefaults.buttonColors(containerColor = Color(0xFF0074cc)),
-                    shape = RoundedCornerShape(20.dp),
+                    onClick = { cheatTerminalOpen = !cheatTerminalOpen },
+                    colors = ButtonDefaults.buttonColors(containerColor = Color(0xFF3933cc)),
                     modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(horizontal = 8.dp)
+                        .width(140.dp)
+                        .height(48.dp)
                 ) {
-                    Text("Exit", color = Color.White)
+                    Text(
+                        if (cheatTerminalOpen) "Close Terminal" else "Open Terminal",
+                        fontSize = 12.sp
+                    )
+                }
+
+                // Chat toggle
+                Button(
+                    onClick = { chatOpen = !chatOpen },
+                    colors = ButtonDefaults.buttonColors(containerColor = Color(0xFF0074cc)),
+                    modifier = Modifier
+                        .width(140.dp)
+                        .height(48.dp)
+                ) {
+                    Text(
+                        if (chatOpen) "Close Chat" else "Open Chat",
+                        fontSize = 14.sp
+                    )
                 }
             }
-        },
-        dismissButton = {}
-    )
-}
-
-
-@Composable
-fun DiceRollingButton(
-    text: String,
-    color: Color,
-    onClick: () -> Unit,
-    diceValue: Int?,
-    enabled: Boolean = true,
-    onRollComplete: () -> Unit = {}
-) {
-
-    var isPressed by remember { mutableStateOf(false) }
-    var rotateAngle by remember { mutableFloatStateOf(0f) }
-
-    val rotation by animateFloatAsState(
-        targetValue = rotateAngle,
-        animationSpec = tween(durationMillis = 1000, easing = FastOutSlowInEasing)
-    )
-    val scale by animateFloatAsState(if (isPressed && enabled) 1.1f else 1f, animationSpec = tween(150))
-    val buttonColor by animateColorAsState(
-        targetValue = when {
-            !enabled      -> Color.Gray
-            isPressed     -> color.copy(alpha = 0.7f)
-            else          -> color
-        },
-        animationSpec = tween(durationMillis = 150)
-    )
-
-    Button(
-        onClick = {
-            if (!enabled) return@Button
-            isPressed = true
-            rotateAngle += 720f
-            onClick()
-            onRollComplete()
-        },
-        enabled = enabled,
-        modifier = Modifier.height(56.dp).scale(scale).rotate(rotation),
-        shape = RoundedCornerShape(8.dp),
-        colors = ButtonDefaults.buttonColors(containerColor = buttonColor)
-    ) {
-        Text(text, fontSize = 18.sp)
-    }
-
-    Spacer(modifier = Modifier.height(16.dp))
-
-    // show the face
-    DiceFace(diceValue)
-
-    LaunchedEffect(isPressed) {
-        if (isPressed) {
-            delay(1000)
-            isPressed = false
         }
-    }
-}
 
-@Composable
-fun DiceFace(diceValue: Int?) {
-    Box(
-        modifier = Modifier
-            .size(100.dp)
-            .background(Color.White, RoundedCornerShape(12.dp))
-            .padding(10.dp),
-        contentAlignment = Alignment.Center
-    ) {
-        Text(
-            text = diceValue?.toString() ?: "?",
-            fontSize = 36.sp,
-            fontWeight = FontWeight.Bold,
-            color = Color.Black
-        )
+        // Chat Overlay
+
+        if (chatOpen) {
+            ChatOverlay(
+                chatMessages = chatMessages,
+                chatInput = chatInput,
+                onChatInputChange = { chatInput = it },
+                currentPlayerId = currentPlayerId,
+                onSendMessage = {
+                    if (chatInput.isNotBlank()) {
+                        webSocketClient.logic().sendChatMessage(currentPlayerId, chatInput)
+                        chatInput = ""
+                    }
+                },
+                playerColorMap = playerColorMap,
+                onClose = { chatOpen = false }
+            )
+        }
+
+        // DEALS
+
+        if (showDealDialog) {
+            DealDialog(
+                players = players.filter { it.id != localPlayerId },
+                senderId = localPlayerId,
+                allProperties = properties,
+                receiver = selectedReceiver,
+                avatarMap = avatarMap,
+                onReceiverChange = { selectedReceiver = it },
+                onSendDeal = { deal ->
+                    val json = Gson().toJson(deal)
+                    webSocketClient.sendMessage(json)
+                    showDealDialog = false
+                    selectedReceiver = null
+                },
+                onDismiss = {
+                    showDealDialog = false
+                    selectedReceiver = null
+                }
+            )
+        }
+
+        if (showIncomingDialog && incomingDeal != null) {
+            val receiverProps = properties.filter { it.ownerId == localPlayerId }.map { it.id }
+
+            IncomingDealDialog(
+                proposal = incomingDeal,
+                senderName = players.find { it.id == incomingDeal.fromPlayerId }?.name ?: "???",
+                allProperties = properties,
+                receiverProperties = receiverProps,
+                isMyTurn = isMyTurn,
+                onAccept = {
+                    val response = DealResponseMessage(
+                        type = "DEAL_RESPONSE",
+                        fromPlayerId = localPlayerId,
+                        toPlayerId = incomingDeal.fromPlayerId,
+                        responseType = DealResponseType.ACCEPT,
+                        counterPropertyIds = listOf(),
+                        counterMoney = 0
+                    )
+                    webSocketClient.sendMessage(Gson().toJson(response))
+                    setShowIncomingDialog(false)
+                    setIncomingDeal(null)
+                },
+                onDecline = {
+                    val response = DealResponseMessage(
+                        type = "DEAL_RESPONSE",
+                        fromPlayerId = localPlayerId,
+                        toPlayerId = incomingDeal.fromPlayerId,
+                        responseType = DealResponseType.DECLINE,
+                        counterPropertyIds = listOf(),
+                        counterMoney = 0
+                    )
+                    webSocketClient.sendMessage(Gson().toJson(response))
+                    setShowIncomingDialog(false)
+                    setIncomingDeal(null)
+                },
+                onCounter = {
+                    isCountering = true
+                    setShowIncomingDialog(false)
+                }
+            )
+        }
+
+        if (isCountering && incomingDeal != null) {
+            DealDialog(
+                players = players.filter { it.id != localPlayerId },
+                senderId = localPlayerId,
+                allProperties = properties,
+                receiver = players.find { it.id == incomingDeal.fromPlayerId },
+                avatarMap = avatarMap,
+                initialRequested = incomingDeal.offeredPropertyIds,
+                initialOffered = incomingDeal.requestedPropertyIds,
+                initialMoney = incomingDeal.offeredMoney,
+                onSendDeal = { counter ->
+                    val counterProposal = DealProposalMessage(
+                        type = "DEAL_PROPOSAL",
+                        fromPlayerId = localPlayerId,
+                        toPlayerId = incomingDeal.fromPlayerId,
+                        offeredPropertyIds = counter.requestedPropertyIds,
+                        requestedPropertyIds = counter.offeredPropertyIds,
+                        offeredMoney = counter.offeredMoney
+                    )
+                    webSocketClient.sendMessage(Gson().toJson(counterProposal))
+                    isCountering = false
+                    setIncomingDeal(null)
+                },
+                onDismiss = {
+                    isCountering = false
+                    setIncomingDeal(null)
+                }
+            )
+        }
+
+        // Cheat Terminal Overview
+        if (cheatTerminalOpen) {
+            CheatTerminalOverlay(
+                cheatMessages = cheatMessages,
+                cheatInput = cheatInput,
+                onCheatInputChange = { cheatInput = it },
+                onSendCheat = {
+                    if (cheatInput.isNotBlank()) {
+                        webSocketClient.logic().sendCheatMessage(currentPlayerId, cheatInput)
+                        cheatInput = ""
+                    }
+                },
+                onClose = { cheatTerminalOpen = false }
+            )
+        }
+        // Popup for CHANCE and COMMUNITY_CHEST
+        if (drawnCardType != null && drawnCardId != null) {
+            CardPopup(
+                cardType = drawnCardType,
+                cardId = drawnCardId,
+                cardDesc = drawnCardDesc,
+                onDismiss = onCardDialogDismiss
+            )
+        }
     }
 }
